requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
playwright>=1.40.0
aiohttp>=3.9.0
aiosqlite>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.9
watchfiles>=0.21.0
httpx>=0.26.0
rendercv>=2.3
ruamel.yaml>=0.18.0
pypdf>=5.0.0
supabase>=2.4.0
PyJWT>=2.8.0
openai>=1.12.0
beautifulsoup4>=4.12.0
pydantic[email]>=2.0.0
